fix: promotion page refactor
